requirements.txt
apscheduler==3.11.0
bcrypt==4.2.1

# dspy
dspy-ai==2.6.17

# spacy model download
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz#egg=en_core_web_sm

# app dependencies
fastapi[standard]==0.115.11
httpx==0.28.1

# auth
pympler==1.1
python-multipart==0.0.18
rich==13.9.4

# util libraries
blis==0.7.11
spacy==3.7.5
uvicorn[standard]==0.29.0

# weaviate
weaviate-client==4.10.4
